backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
minio==7.2.0
litellm==1.17.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
setuptools==69.0.0

